Use Case Diagram
TaskID: HDB-7
Please make the Use Case diagram for the project.

The diagram shall have the use cases for the application (the
application is in continuous development so it would suffer changes
to the Use Case diagram as the project evolves).

The diagram shall be made in StarUML

The diagram shall be attached in the ticket after it’s completion

Type:  Feature
Author: Haragis Horea

Description:
Use Case diagram

Files:
Design/Use_Case_Diagram/