Frontend folder
TaskID:  HDB-13
Please make the design of the UI for the application.

The UI shall have a modern look and satisfy all the functionalities
that were agreed on.

Multiple subtasks for this one can be created by you for better
traceability. Talk to your colleague and agree on the workload
distribution.
Type: Feature
Author: Rad Antonio

Description:
Frontend folder

Files:
frontend/